Fix balance by count
The balance by count module could often pick the wrong partition to change
since it goes in numeric partition order. A case for this is the following
replica layout for a topic with 4 partitions:
[4, 3], [2, 4], [2, 1], [3,1]

On the first pass, the code will see replica 2 in position 0 as a problem,
and it will change the first partition it finds, partition 2. Since
replica 4 has a problem at position 1, it will swap the replica positions.
At partition 3, it will see that both replicas have problems again, and
swap the replica positions. On the last partition, it will find no problem
with either replica, and finish.

What actually needed to happen was just partition 3 needed to have the
replica positions swapped. A second pass over this topic will cause
partition 2 to be swapped back, leaving the topic in a good state.