Add rate balance module
The rate balance module is a variant on size balance, except that it scales the size used based on the retention of the topic. Topics that are retained longer
than the cluster default (now specified on the command line with --default-retention) will be weighted lower, such that if a topic has double the retention of
the cluster, its scaled size will be half its actual size on disk. This will more closely approximate data rate balancing, which was the original intention of
the size module.