Cluster client (#69)
The existing kafka-python library has a lack of features around admin operations. This is specifically around ops like fetching topic and group information, and fetching and committing offsets. This PR adds a full client around the cluster that allows those operations. It is thread-safe, and parallelizes the broker operations to make it a little bit faster.

This uses the underlying protocol definitions, but it only provides a few operations that are safe for use in the cluster. It does support both PLAINTEXT and SSL operation, and supports client certificate operation.