Fix protocol definition for bytes (#71)
There's a flaw in the protocol implementation where bytes are defined as having an int16 length. This is incorrect - it's int32.

In addition, it was found while fixing this that the schema for member assignments in a group was wrong, as it contains an array of partitions, not just a raw int32 value.