Handle case where ListOffset returns an empty array of offsets
There was an assumption made that ListOffset always returns at least one offset. However, in the case where the timestamp requested is out of range (or for brokers that do not have a time index, when the offset is in the first segment) where an empty array is returned. This updates the TopicOffsets model to handle that case and insert a -1 instead to indicate that no offset was available. I also added a test specifically for that case.